+++ pkg/client/client.go
@@ -1,4 +1,4 @@
-package mariadb
+package client
 
 import (
 	"bytes"
